Update target framework and dependencies
Upgraded project's target framework from .NET 7.0 to .NET 8.0 and updated dependencies to their latest versions. Reason for this change is to get the benefits of the new features and improvements in .NET 8.0 and the updated packages. For instance, the updated Mapster and MimeKit libraries provide better performance and various bug fixes.