Simplify service registration and remove redundant logging
This commit refactors the configuration of services in Program.cs for cleaner code. The MassTransit service registration was moved into the fluent API call to make it more concise. In EmailCommandConsumer.cs, an unnecessary JSON serialization log was removed due to redundancy as it was causing duplicate logs in the console output. This cleans up the log output shown in testing and production environments.