"Remove unused output cache and improve user photo management"
Removed the output caching middleware from the application startup as it was no longer necessary. This was decided because it was causing unnecessary overhead.

Clean-up of unused namespaces in several test files was also done.

Adjusted the profile photo handling in UsersController. Replaced the 'Ok' response of UploadProfilePhoto action with 'NoContent'.
Moved the deletion of existing photos in the upload directory to the top of the method to ensure that the old photos are deleted before the new one is uploaded.

In the GetProfilePhoto action, made it possible to retrieve photos of any user by their id, not just the current user. Changed the default content-type fallback to use a constant from System.Net.Mime.MediaTypeNames.Application.Octet.

In the GetPhoto helper method, changed the logic to check if the file path of the photo is null instead of checking if the files array length is 0. Kept file reading in a separate call to make code more readable.