Remove test-compose workflow and update dotnet workflow
The file .github/workflows/test-compose.yml was deleted as it was no longer necessary. The 'dotnet.yml' workflow file was updated to change the trigger branches from "docker/with-github-actions" to "master". The task to build Docker image was also removed because it's not necessary in the current scope of project. The steps to build and test the .NET project remain. This simplifies our CI/CD process and limits actions to the master branch only.