Docker/with GitHub actions -test pipelines (#9)
* Docker works good - initial step

* Added limits to docker-compose

* ALl works

* Refactor configuration for RabbitMQ and implement logging with Serilog

RabbitMQ configuration has been refactored to be driven by appsettings.json for more flexibility across environments. Logging infrastructure has been implemented with the Serilog library for both the Web and EmailSender projects. Corresponding settings have been added to the appsettings.json files. Health checks were adjusted in docker-compose file. Also, log consumption is added in the EmailCommandConsumer, which will help to track the email command processing.

* Create dotnet.yml for building dotnet

* Small change to check action

* Build docker image

* Create test-compose.yml

* Added newline at end of CurrentUser.cs file

A newline was added at the end of the CurrentUser.cs file to comply with coding standards and conventions. This change does not affect the functionality of the code.

* Remove test-compose workflow and update dotnet workflow

The file .github/workflows/test-compose.yml was deleted as it was no longer necessary. The 'dotnet.yml' workflow file was updated to change the trigger branches from "docker/with-github-actions" to "master". The task to build Docker image was also removed because it's not necessary in the current scope of project. The steps to build and test the .NET project remain. This simplifies our CI/CD process and limits actions to the master branch only.

* Added cancellation token

---------

Co-authored-by: Alexey Kornienko <alexey.kornienko@smartodds.co.uk>
Co-authored-by: Oleksii Korniienko <oleksii.korniienko@idealscorp.com>