"Refactored CORS policy and migration process" (#21)
In this commit, the code was restructured to centralize the CORS policy and the database migration process. The CORS policy has been moved to a constants class which will help to keep configurations consistent across the application. The database migration has been moved to a MigrationHostedService, allowing the migration process to occur when the application starts, rather than on each request. This can increase the application's performance by removing potentially unnecessary database operations from the request pipeline.

Co-authored-by: Oleksii Korniienko <oleksii.korniienko@idealscorp.com>