Refactor version trim code and cleanup readme markdown
Streamlined version trimming code in TwitPoster.Shared.yml workflow file for better readability and maintainability. The trim operation is now included in the Write-Host command, which improves clarity.

Also, an extra line was removed from readme.md file to ensure consistence in document spacing. This cleanup brings a tidy look to the documentation.