Update version trim process and readme file
This commit changes the method of how we were reducing leading and trailing spaces from the version number. Past method was problematic when encountered unnecessary white spaces in the middle of strings. It now split and join method to handle all kinds of white spaces. Additionally, two line breaks were added to the readme file for better readability.