Update build and test orders in workflow file
This commit rearranges the order of build and test commands in the TwitPoster.Shared.yml GitHub workflow file. The original order was restore dependencies, calculate version, create a tag if needed, build, and test. The updated order is restore dependencies, build, test, calculate version, and create a tag if needed. The dotnet pack command was slightly