Add deploy step to TwitPoster.Shared workflow
The commit includes addition of the 'deploy' step in the GitHub Actions workflow for the TwitPoster.Shared project. The deploy step is essential to allow building, testing, and packaging of the TwitPoster.Shared project. It starts by fetching all git history, setting the tagPrefix, calculating version, creating tags, packing, pushing the new version