Update Services for Production
Updated the services to run on the production environment. Removed the condition for using RabbitMq only on the development environment in Program.cs, as it is now required to run on the production environment as well. Also, commented out caching functionality in TestServiceBusService.cs as it creates unnecessary overhead in production.