Add new origin and update Redis port
The CORS policy in `Program.cs` file is updated to allow requests from another origin "https://wallyrion.github.io". This change is necessary to connect to the front-end application hosted on Github Pages. The Redis port in `appsettings.Development.json` is updated from 6380 to 6379 due to changes in the local dev environment.