Update TwitPoster workflow for dev branch and separate build and deployment
The TwitPoster GitHub Actions workflow has been modified to also respond to pushes and pull requests on the 'dev' branch in addition to the 'master' branch. This change allows for continuous integration and delivery on the development branch as well. Moreover, the 'build' and 'deploy' jobs have been separated for a more flexible and clear workflow. Lastly, an artifact upload/download step has been added to complement the separate jobs. These changes enhance automated