Switch messaging service to Azure Service Bus
The messaging service is switched from local RabbitMQ service to Azure Service Bus. This is done to unify communication between applications and to make use of Azure’s powerful services for production environments. Changes made include the addition of Azure Bus configuration and removal of RabbitMQ configuration. Additionally, MailOptions, IServiceBusMessageScheduler and EmailCommandConsumer are added to accommodate for these changes.