Add Location functionality to the TwitPoster app
A new feature was added to the TwitPoster application that allows for the retrieval of location data (country, state, and city). New code includes the creation of a LocationController in the Web Project, a LocationService in the BLL Project, and the corresponding DTOs and interfaces. This functionality makes it possible for a user to retrieve location-specific data for their posts. Additionally, caching was implemented to minimize external requests and improve performance and efficiency. The application can now be configured to use either Redis or in-memory caching through the appsettings file.