Add null check for post in LikePostTests
Added a null check for 'post' variable in LikePostTests to avoid NullReferenceException. This check ensures that 'post' is not null before trying to access its 'Id' property. This action is performed because we are directly accessing the 'Id' from the 'post'. This check will prevent the system from throwing an error when 'post' is null.