Refactor service constructors to use less code (#44)
* Refactor service constructors to use less code

Services constructors were refactored to reduce redundant code. The changes involve simplification of constructors in `MemoryCacheService`, `LocationClient`, and `LocationService` by using injected parameters directly, instead of assigning them to private fields. This makes the code cleaner and more efficient.

* Add null check for post in LikePostTests

Added a null check for 'post' variable in LikePostTests to avoid NullReferenceException. This check ensures that 'post' is not null before trying to access its 'Id' property. This action is performed because we are directly accessing the 'Id' from the 'post'. This check will prevent the system from throwing an error when 'post' is null.

---------

Co-authored-by: Oleksii Korniienko <oleksii.korniienko@idealscorp.com>