Add Docker publishing to TwitPoster workflow
Added an extra step to the Github Actions workflow for the TwitPoster project. After the testing step, Docker Buildx is set up and used to log into the Github container registry. This is followed by a publish step, which builds and pushes a Docker image of the project. This will facilitate the deployment