"Remove unnecessary pull_request trigger in workflow file
The pull_request trigger was removed from the TwitPoster.EmailSender.deploy workflow file. This trigger was redundant as it was also specified on push to master. The removal streamlines the workflow and makes the deployment process on Azure more efficient by triggering a webhook after a successful deployment. This enhances automation and the continuous integration and deployment processes."